% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCodebook.R, R/mbCB.R
\name{createCodebook}
\alias{createCodebook}
\title{Creates a codebook}
\usage{
createCodebook(df, outname = "test", savein = "examples",
  overwrite = T, reportTitle = "My Codebook", output = "html")

createCodebook(df, outname = "test", savein = "examples",
  overwrite = T, reportTitle = "My Codebook", output = "html")
}
\arguments{
\item{df}{Dataframe from which to construct the codebook}

\item{outname}{Base name to be used for output codebook files}

\item{overwrite}{if TRUE (default), overwrites existing files}

\item{reportTitle}{String to be passed to Codebook as title}

\item{outputName}{Name to be used for output codebook files}

\item{saveLocation}{Location to store created codebooks}

\item{outputType}{Type of file to create (currently only HTML)}

\item{saveRmd}{if TRUE (default), keeps intermediary markdown file.  Otherwise deletes}

\item{df}{Dataframe from which to construct the codebook}

\item{overwrite}{if TRUE (default), overwrites existing files}

\item{reportTitle}{String to be passed to Codebook as title}
}
\description{
Creates a codebook

Creates a codebook
}
