% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDataFrame.R
\name{summarizeDataFrame}
\alias{summarizeDataFrame}
\title{Creates a table summarizing key data frame dimensions}
\usage{
summarizeDataFrame(df)
}
\arguments{
\item{df}{that will be summarized}
}
\value{
A dataframe with summary statistics
}
\description{
Creates a table summarizing key data frame dimensions
}
