% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeVariable.R
\name{summarizeVariable}
\alias{summarizeVariable}
\title{Produces summary table of  variable}
\usage{
summarizeVariable(var, varClass)
}
\arguments{
\item{var}{Variable to be summarized}

\item{varClass}{Variable Class to be summarized}
}
\value{
Dataframe with summary information
}
\description{
Produces summary table of  variable
}
