% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeVariables.R
\name{summarizeVariables}
\alias{summarizeVariables}
\title{Produces sumamry table of all variables in dataframe}
\usage{
summarizeVariables(df)
}
\arguments{
\item{df}{Dataframe containing variables to be summarized}
}
\value{
Dataframe with summary information
}
\description{
Produces sumamry table of all variables in dataframe
}
