% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableOverview.R
\name{variableOverview}
\alias{variableOverview}
\title{creates variable overview showing type, no. missing, etc.}
\usage{
variableOverview(var)
}
\arguments{
\item{var}{Variable to be summarized}
}
\value{
Dataframe with overview information
}
\description{
creates variable overview showing type, no. missing, etc.
}
