% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCodebook.R
\name{createCodebook}
\alias{createCodebook}
\title{Creates a codebook}
\usage{
createCodebook(df, outputName = "codebook", saveLocation = "temp/",
  overwrite = T, reportTitle = "My Codebook", outputType = "html",
  saveRmd = TRUE)
}
\arguments{
\item{df}{Dataframe from which to construct the codebook}

\item{outputName}{Name to be used for output codebook files}

\item{saveLocation}{Location to store created codebooks}

\item{overwrite}{if TRUE (default), overwrites existing files}

\item{reportTitle}{String to be passed to Codebook as title}

\item{outputType}{Type of file to create (currently only HTML)}

\item{saveRmd}{if TRUE (default), keeps intermediary markdown file.  Otherwise deletes}
}
\description{
Creates a codebook
}
